# Project_1/micro-loan-pwa/openapi.yaml
openapi: 3.0.3
info:
  title: Micro-Loan Calculator API
  description: |
    Backend API for the Micro-Loan Calculator PWA.
    
    Supports:
    - Loan calculations with validation
    - Cloud sync for loan history and savings goals
    - Financial literacy content delivery
    - User authentication (optional)
    - Lender database (future)
    
    PRD Reference: Sections 3.1-3.4, Phase 3 requirements
  version: 1.0.0
  contact:
    name: API Support
    email: support@microloan.app
servers:
  - url: http://localhost:3001/api
    description: Development server
  - url: https://api.microloan.app/api
    description: Production server
...
paths:
...
  /savings:
    get:
      tags:
        - Savings
      summary: Get all savings goals
      description: Retrieve all savings goals for the authenticated user
      operationId: getSavingsGoals
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of savings goals
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SavingsGoalsResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      tags:
        - Savings
      summary: Create a savings goal
      operationId: createSavingsGoal
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSavingsGoalRequest'
      responses:
        '201':
          description: Savings goal created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SavingsGoal'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /savings/{id}:
    get:
      tags:
        - Savings
      summary: Get a specific savings goal
      operationId: getSavingsGoalById
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Savings goal details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SavingsGoal'
        '404':
          description: Goal not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      tags:
        - Savings
      summary: Update a savings goal
      operationId: updateSavingsGoal
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSavingsGoalRequest'
      responses:
        '200':
          description: Savings goal updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SavingsGoal'
        '404':
          description: Goal not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags:
        - Savings
      summary: Delete a savings goal
      operationId: deleteSavingsGoal
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Goal deleted successfully
        '404':
          description: Goal not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /savings/{id}/transactions:
    post:
      tags:
        - Savings
      summary: Add a transaction to a savings goal
      operationId: addTransaction
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddTransactionRequest'
      responses:
        '201':
          description: Transaction added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transaction'
        '404':
          description: Goal not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SavingsGoal:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        name:
          type: string
          example: Emergency Fund
        targetAmount:
          type: number
          example: 5000
        currentAmount:
          type: number
          example: 1250
        currency:
          type: string
          example: USD
        targetDate:
          type: string
          format: date-time
          nullable: true
        transactions:
          type: array
          items:
            $ref: '#/components/schemas/Transaction'
        createdAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
          nullable: true
    Transaction:
      type: object
      properties:
        id:
          type: string
          format: uuid
        amount:
          type: number
          description: Positive for deposit, negative for withdrawal
          example: 100
        type:
          type: string
          enum: [deposit, withdrawal]
        note:
          type: string
          nullable: true
        date:
          type: string
          format: date-time
    CreateSavingsGoalRequest:
      type: object
      required:
        - name
        - targetAmount
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        targetAmount:
          type: number
          minimum: 1
        currency:
          type: string
          default: USD
        targetDate:
          type: string
          format: date-time
          nullable: true
        localId:
          type: string
          description: Client-side ID for offline sync
    UpdateSavingsGoalRequest:
      type: object
      properties:
        name:
          type: string
        targetAmount:
          type: number
        targetDate:
          type: string
          format: date-time
          nullable: true
    AddTransactionRequest:
      type: object
      required:
        - amount
      properties:
        amount:
          type: number
          description: Positive for deposit, negative for withdrawal
        note:
          type: string
          maxLength: 200
        date:
          type: string
          format: date-time
          description: Defaults to current time if not provided
        localId:
          type: string
    SavingsGoalsResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: '#/components/schemas/SavingsGoal'

    # ==================
    # LESSON SCHEMAS
    # ==================
...
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              example: VALIDATION_ERROR
            message:
              type: string
              example: Invalid input data
            details:
              type: object
              additionalProperties: true
...